+++ web/themes/custom/itkdev/itkdev_project_theme/js/disableAutoEvaluation.js
@@ -1,0 +1,7 @@
+let checkbox = document.getElementById('edit-project-track-evaluation-set-manual')
+let autoEvaluation = document.getElementById('auto-evaluation')
+
+const updateState = () => autoEvaluation.classList.toggle('text-slate-300', checkbox.checked)
+checkbox.addEventListener("change", updateState)
+// Initialize.
+updateState()