+++ web/themes/custom/itkdev/itkdev_project_theme/tailwind.config.js
@@ -11,6 +11,12 @@ module.exports = {
     "bg-slate-300",
     "bg-yellow-700",
     "bg-red-700",
-    "grid-cols-4"
+    "grid-cols-4",
+    "order-last",
+    "text-slate-300",
+    "border-zinc-400",
+    "border-red-600",
+    "border-yellow-600",
+    "border-green-600",
   ],
 };